% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage_trend.R
\name{trend}
\alias{trend}
\title{This is a sample function}
\usage{
trend(type, provs = c("ON", "BC"), positions = c("Data_entry_clerks"))
}
\arguments{
\item{type}{A choice using "overall" or "specific"}

\item{provs}{a value or vector about province}

\item{positions}{a value or vector about job type}
}
\value{
One or two plot
}
\description{
This is a sample function
}
\examples{
trend("overall")
}
